+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add check on `cm get policies` if the platform is RHACM
 - Add `Standby` column in the `cm get clusterpool`
 - Display an error if the clusterclaim has no namespace,  this can happen when new clusterclaim with the same name is created just of the deletion of a clusterclaim with the same name. [When clusterclaim not yet running, cm use/run generates an error](https://github.com/stolostron/cm-cli/issues/167)
+- Check if Identity Configuration Management is installed when running `cm create authrealm` 
 ## Breaking changes
 
 ## Bug fixes